# Enhance Solving Projects and Problems - Math Quiz - Rock-Paper-Scissors - Problem26.cpp - Problem27.cpp

+++ src/_3_problems_from_21_to_30/_3_6_problem_26/Problem26.cpp
@@ -37,11 +37,11 @@ void printNumbers(
     }
 }
 
-long numbersSum(
+long long numbersSum(
     const short NUMBERS[],
     int numberCount
 ) {
-    long sum = 0;
+    long long sum = 0;
     while (numberCount--)
         sum += NUMBERS[numberCount];
     return sum;
